model/src/detector.py
"""
Deepfake Detection Model Interface
Independent of any database or backend dependencies
"""

import os
import numpy as np
from typing import Dict, Any, Optional, Tuple
from pathlib import Path
import cv2
from PIL import Image
import tensorflow as tf
from dataclasses import dataclass
from enum import Enum

class MediaType(str, Enum):
    IMAGE = "image"
    VIDEO = "video"

class DetectionResult(str, Enum):
    REAL = "real"
    FAKE = "fake"
    UNCERTAIN = "uncertain"

@dataclass
class DetectionOutput:
    """Output structure for detection results"""
    result: DetectionResult
    confidence_score: float
    processing_time_seconds: float
    metadata: Dict[str, Any]
    model_version: str

class DeepfakeDetector:
    """
    Standalone deepfake detection model
    No database or external API dependencies
    """
    
    def __init__(self, model_path: str, model_version: str = "v1.0"):
        self.model_path = Path(model_path)
        self.model_version = model_version
        self.model = None
        self.is_loaded = False
        
    def load_model(self) -> bool:
        """Load the trained model from file"""
        try:
            if self.model_path.exists():
                # Load your trained model here
                # self.model = tf.keras.models.load_model(str(self.model_path))
                # For now, we'll simulate a loaded model
                self.model = "loaded_model_placeholder"
                self.is_loaded = True
                print(f"Model loaded successfully from {self.model_path}")
                return True
            else:
                print(f"Model file not found: {self.model_path}")
                return False
        except Exception as e:
            print(f"Error loading model: {str(e)}")
            return False
    
    def preprocess_image(self, image_path: str) -> Optional[np.ndarray]:
        """Preprocess image for model input"""
        try:
            # Load and preprocess image
            image = cv2.imread(image_path)
            if image is None:
                return None
                
            # Resize to model input size (adjust based on your model)
            image = cv2.resize(image, (224, 224))
            image = cv2.cvtColor(image, cv2.COLOR_BGR2RGB)
            
            # Normalize pixel values
            image = image.astype(np.float32) / 255.0
            
            # Add batch dimension
            image = np.expand_dims(image, axis=0)
            
            return image
        except Exception as e:
            print(f"Error preprocessing image: {str(e)}")
            return None
    
    def preprocess_video(self, video_path: str, max_frames: int = 30) -> Optional[np.ndarray]:
        """Extract and preprocess frames from video"""
        try:
            cap = cv2.VideoCapture(video_path)
            frames = []
            frame_count = 0
            
            while cap.read()[0] and frame_count < max_frames:
                ret, frame = cap.read()
                if not ret:
                    break
                    
                # Resize and preprocess frame
                frame = cv2.resize(frame, (224, 224))
                frame = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
                frame = frame.astype(np.float32) / 255.0
                frames.append(frame)
                frame_count += 1
            
            cap.release()
            
            if frames:
                return np.array(frames)
            return None
            
        except Exception as e:
            print(f"Error preprocessing video: {str(e)}")
            return None
    
    def predict_image(self, image_path: str) -> DetectionOutput:
        """Predict if an image is a deepfake"""
        import time
        start_time = time.time()
        
        if not self.is_loaded:
            if not self.load_model():
                raise RuntimeError("Model not loaded")
        
        # Preprocess image
        processed_image = self.preprocess_image(image_path)
        if processed_image is None:
            raise ValueError("Failed to preprocess image")
        
        # Make prediction (placeholder - replace with actual model inference)
        # prediction = self.model.predict(processed_image)
        # For demonstration, we'll return a mock prediction
        confidence_score = np.random.uniform(0.3, 0.9)  # Replace with actual prediction
        
        # Determine result based on confidence
        if confidence_score > 0.7:
            result = DetectionResult.FAKE
        elif confidence_score < 0.3:
            result = DetectionResult.REAL
        else:
            result = DetectionResult.UNCERTAIN
        
        processing_time = time.time() - start_time
        
        return DetectionOutput(
            result=result,
            confidence_score=confidence_score,
            processing_time_seconds=processing_time,
            metadata={
                "input_type": "image",
                "image_path": image_path,
                "image_shape": processed_image.shape if processed_image is not None else None
            },
            model_version=self.model_version
        )
    
    def predict_video(self, video_path: str) -> DetectionOutput:
        """Predict if a video contains deepfakes"""
        import time
        start_time = time.time()
        
        if not self.is_loaded:
            if not self.load_model():
                raise RuntimeError("Model not loaded")
        
        # Preprocess video
        processed_frames = self.preprocess_video(video_path)
        if processed_frames is None:
            raise ValueError("Failed to preprocess video")
        
        # Make predictions on frames (placeholder - replace with actual model inference)
        # predictions = [self.model.predict(np.expand_dims(frame, axis=0)) for frame in processed_frames]
        # For demonstration, we'll return a mock prediction
        frame_predictions = [np.random.uniform(0.2, 0.8) for _ in processed_frames]
        
        # Aggregate predictions (you might want to use different strategies)
        confidence_score = np.mean(frame_predictions)
        
        # Determine result based on confidence
        if confidence_score > 0.6:
            result = DetectionResult.FAKE
        elif confidence_score < 0.4:
            result = DetectionResult.REAL
        else:
            result = DetectionResult.UNCERTAIN
        
        processing_time = time.time() - start_time
        
        return DetectionOutput(
            result=result,
            confidence_score=confidence_score,
            processing_time_seconds=processing_time,
            metadata={
                "input_type": "video",
                "video_path": video_path,
                "frames_analyzed": len(processed_frames),
                "frame_predictions": frame_predictions
            },
            model_version=self.model_version
        )
    
    def detect(self, media_path: str, media_type: MediaType) -> DetectionOutput:
        """Universal detection method"""
        if media_type == MediaType.IMAGE:
            return self.predict_image(media_path)
        elif media_type == MediaType.VIDEO:
            return self.predict_video(media_path)
        else:
            raise ValueError(f"Unsupported media type: {media_type}")


# Factory function for easy instantiation
def create_detector(model_path: str, model_version: str = "v1.0") -> DeepfakeDetector:
    """Create and return a DeepfakeDetector instance"""
    return DeepfakeDetector(model_path, model_version)
